requirements.txt
deta==1.1.0
pandas==2.0.1
plotly==5.14.1
python-dotenv==1.0.0
PyYAML==6.0
PyYAML==6.0
streamlit==1.22.0
streamlit_authenticator==0.2.2
streamlit_option_menu==0.3.4
